requirements.txt
matplotlib==3.10.0
numpy==2.2.2
oracledb>=1.4.0
psycopg==3.2.6
pylint==3.3.3
pyyaml>=6.0.1
sqlparse>=0.4.4
